accessible/tests/mochitest/text/test_singleline.html

<!DOCTYPE html>
<html>
<head>
  <title>nsIAccessibleText getText related function tests for html:input,html:div and html:textarea</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>
  <script type="application/javascript">
    if (navigator.platform.startsWith("Mac")) {
      SimpleTest.expectAssertions(0, 14);
    } else {
      SimpleTest.expectAssertions(14);
    }

    function doTest()
    {
      // __h__e__l__l__o__ __m__y__ __f__r__i__e__n__d__
      //  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15

      ////////////////////////////////////////////////////////////////////////
      // characterCount

      testCharacterCount(["input", "div"], 15);
      testCharacterCount(["textarea"], 16);

      ////////////////////////////////////////////////////////////////////////
      // getText

      var IDs = ["input", "div", "textarea"];
      testText(IDs, 0, 1, "h");
      testText(IDs, 1, 3, "el");
      testText(IDs, 14, 15, "d");
      testText(IDs, 0, 15, "hello my friend");

      ////////////////////////////////////////////////////////////////////////
      // getTextAfterOffset

      var IDs = [ "input", "div", "editable", "textarea" ];
      var regularIDs = [ "input", "div", "editable" ];

      // BOUNDARY_CHAR

      testCharAfterOffset(IDs, 0, "e", 1, 2);
      testCharAfterOffset(IDs, 1, "l", 2, 3);
      testCharAfterOffset(regularIDs, 14, "", 15, 15);
      testCharAfterOffset("textarea", 14, "\n", 15, 16);

      // XXX: why are 15/15 expected? there's no 16 offset we are trying to
      // get an offset for?
      testTextAfterOffset(15, BOUNDARY_CHAR, "", 15, 15,
			  "input", kOk, kTodo, kTodo,
			  "div", kOk, kTodo, kTodo,
			  "editable", kOk, kTodo, kTodo);
      testCharAfterOffset("textarea", 15, "", 16, 16);

      // BOUNDARY_WORD_START
      testTextAfterOffset(0, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAfterOffset(1, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAfterOffset(5, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAfterOffset(6, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAfterOffset(7, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAfterOffset(8, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAfterOffset(9, BOUNDARY_WORD_START, "", 15, 15, IDs);
      testTextAfterOffset(11, BOUNDARY_WORD_START, "", 15, 15, IDs);
      testTextAfterOffset(14, BOUNDARY_WORD_START, "", 15, 15, IDs);
      testTextAfterOffset(15, BOUNDARY_WORD_START, "", 15, 15, IDs);

      // BOUNDARY_WORD_END
      testTextAfterOffset(0, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAfterOffset(1, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAfterOffset(5, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAfterOffset(6, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAfterOffset(7, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAfterOffset(8, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAfterOffset(9, BOUNDARY_WORD_END, "", 15, 15, IDs);
      testTextAfterOffset(11, BOUNDARY_WORD_END, "", 15, 15, IDs);
      testTextAfterOffset(14, BOUNDARY_WORD_END, "", 15, 15, IDs);
      testTextAfterOffset(15, BOUNDARY_WORD_END, "", 15, 15, IDs);

      // BOUNDARY_LINE_START
      testTextAfterOffset(0, BOUNDARY_LINE_START, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kTodo);
      testTextAfterOffset(1, BOUNDARY_LINE_START, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kTodo);
      testTextAfterOffset(14, BOUNDARY_LINE_START, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kTodo);
      testTextAfterOffset(15, BOUNDARY_LINE_START, "", 15, 15,
                          "input", kOk, kOk, kOk,
                          "div", kOk, kOk, kOk,
                          "editable", kOk, kOk, kOk,
                          "textarea", kTodo, kOk, kTodo);

      // BOUNDARY_LINE_END
      testTextAfterOffset(0, BOUNDARY_LINE_END, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kOk);
      testTextAfterOffset(1, BOUNDARY_LINE_END, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kOk);
      testTextAfterOffset(14, BOUNDARY_LINE_END, "", 15, 15,
                          "input", kTodo, kTodo, kOk,
                          "div", kTodo, kTodo, kOk,
                          "editable", kTodo, kTodo, kOk,
                          "textarea", kTodo, kTodo, kOk);
      testTextAfterOffset(15, BOUNDARY_LINE_END, "", 15, 15,
                          "input", kOk, kTodo, kTodo,
                          "div", kOk, kTodo, kTodo,
                          "editable", kOk, kTodo, kTodo,
                          "textarea", kTodo, kOk, kTodo);

      ////////////////////////////////////////////////////////////////////////
      // getTextBeforeOffset

      var IDs = [ "input", "div", "editable", "textarea" ];

      // BOUNDARY_CHAR
      testCharBeforeOffset(IDs, 0, "", 0, 0);
      testCharBeforeOffset(IDs, 1, "h", 0, 1);
      testCharBeforeOffset(IDs, 14, "n", 13, 14);
      testCharBeforeOffset(IDs, 15, "d", 14, 15);

      // BOUNDARY_WORD_START
      testTextBeforeOffset(0, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(5, BOUNDARY_WORD_START, "", 0, 0, IDs);
      testTextBeforeOffset(6, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextBeforeOffset(7, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextBeforeOffset(8, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextBeforeOffset(9, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextBeforeOffset(10, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextBeforeOffset(14, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextBeforeOffset(15, BOUNDARY_WORD_START, "my ", 6, 9,
                           "input", kOk, kOk, kOk,
                           "div", kOk, kOk, kOk,
                           "editable", kOk, kOk, kOk,
                           "textarea", kTodo, kTodo, kTodo);

      // BOUNDARY_WORD_END
      testTextBeforeOffset(0, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(5, BOUNDARY_WORD_END, "", 0, 0, IDs);
      testTextBeforeOffset(6, BOUNDARY_WORD_END, "hello", 0, 5, IDs);
      testTextBeforeOffset(7, BOUNDARY_WORD_END, "hello", 0, 5, IDs);
      testTextBeforeOffset(8, BOUNDARY_WORD_END, "hello", 0, 5, IDs);
      testTextBeforeOffset(9, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextBeforeOffset(10, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextBeforeOffset(14, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextBeforeOffset(15, BOUNDARY_WORD_END, " my", 5, 8, IDs);

      // BOUNDARY_LINE_START
      testTextBeforeOffset(0, BOUNDARY_LINE_START, "", 0, 0,
                           "input", kOk, kOk, kOk,
                           "div", kOk, kOk, kOk,
                           "editable", kOk, kOk, kOk,
                           "textarea", kOk, kOk, kOk);
      testTextBeforeOffset(1, BOUNDARY_LINE_START, "", 0, 0,
                           "input", kTodo, kOk, kTodo,
                           "div", kTodo, kOk, kTodo,
                           "editable", kTodo, kOk, kTodo,
                           "textarea", kTodo, kOk, kTodo);
      testTextBeforeOffset(14, BOUNDARY_LINE_START, "", 0, 0,
                           "input", kTodo, kOk, kTodo,
                           "div", kTodo, kOk, kTodo,
                           "editable", kTodo, kOk, kTodo,
                           "textarea", kTodo, kOk, kTodo);
      testTextBeforeOffset(15, BOUNDARY_LINE_START, "", 0, 0,
                           "input", kTodo, kOk, kTodo,
                           "div", kTodo, kOk, kTodo,
                           "editable", kTodo, kOk, kTodo,
                           "textarea", kTodo, kOk, kTodo);

      // BOUNDARY_LINE_END
      testTextBeforeOffset(0, BOUNDARY_LINE_END, "", 0, 0,
                           "input", kOk, kOk, kOk,
                           "div", kOk, kOk, kOk,
                           "editable", kOk, kOk, kOk,
                           "textarea", kOk, kOk, kOk);
      testTextBeforeOffset(1, BOUNDARY_LINE_END, "", 0, 0,
                           "input", kTodo, kOk, kTodo,
                           "div", kTodo, kOk, kTodo,
                           "editable", kTodo, kOk, kTodo,
                           "textarea", kTodo, kOk, kTodo);
      testTextBeforeOffset(14, BOUNDARY_LINE_END, "", 0, 0,
                           "input", kTodo, kOk, kTodo,
                           "div", kTodo, kOk, kTodo,
                           "editable", kTodo, kOk, kTodo,
                           "textarea", kTodo, kOk, kTodo);
      testTextBeforeOffset(15, BOUNDARY_LINE_END, "", 0, 0,
                           "input", kOk, kOk, kOk,
                           "div", kOk, kOk, kOk,
                           "editable", kOk, kOk, kOk,
                           "textarea", kTodo, kOk, kTodo);

      ////////////////////////////////////////////////////////////////////////
      // getTextAtOffset

      IDs = [ "input", "div", "editable", "textarea" ];
      regularIDs = [ "input", "div", "editable" ];

      // BOUNDARY_CHAR

      testCharAtOffset(IDs, 0, "h", 0, 1);
      testCharAtOffset(IDs, 1, "e", 1, 2);
      testCharAtOffset(IDs, 14, "d", 14, 15);
      testCharAtOffset(regularIDs, 15, "", 15, 15);
      testCharAtOffset("textarea", 15, "\n", 15, 16);
      testCharAtOffset("textarea", 16, "", 16, 16);

      // BOUNDARY_WORD_START
      testTextAtOffset(0, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextAtOffset(1, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextAtOffset(5, BOUNDARY_WORD_START, "hello ", 0, 6, IDs);
      testTextAtOffset(6, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAtOffset(7, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAtOffset(8, BOUNDARY_WORD_START, "my ", 6, 9, IDs);
      testTextAtOffset(9, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAtOffset(10, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAtOffset(14, BOUNDARY_WORD_START, "friend", 9, 15, IDs);
      testTextAtOffset(15, BOUNDARY_WORD_START, "friend", 9, 15, IDs);

      // BOUNDARY_WORD_END
      testTextAtOffset(0, BOUNDARY_WORD_END, "hello", 0, 5, IDs);
      testTextAtOffset(1, BOUNDARY_WORD_END, "hello", 0, 5, IDs);
      testTextAtOffset(5, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAtOffset(6, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAtOffset(7, BOUNDARY_WORD_END, " my", 5, 8, IDs);
      testTextAtOffset(8, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAtOffset(9, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAtOffset(10, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAtOffset(14, BOUNDARY_WORD_END, " friend", 8, 15, IDs);
      testTextAtOffset(15, BOUNDARY_WORD_END, " friend", 8, 15, IDs);

      // BOUNDARY_LINE_START
      testTextAtOffset(0, BOUNDARY_LINE_START, "hello my friend", 0, 15,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);
      testTextAtOffset(1, BOUNDARY_LINE_START, "hello my friend", 0, 15,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);
      testTextAtOffset(14, BOUNDARY_LINE_START, "hello my friend", 0, 15,
                       "input", kOk, kOk, kOk,
                       "div", kOk, kOk, kOk,
                       "editable", kOk, kOk, kOk,
                       "textarea", kTodo, kOk, kTodo);
      testTextAtOffset(15, BOUNDARY_LINE_START, "hello my friend", 0, 15,
		       "input", kOk, kOk, kOk,
		       "div", kOk, kOk, kOk,
		       "editable", kOk, kOk, kOk,
		       "textarea", kTodo, kOk, kTodo);

      // BOUNDARY_LINE_END
      testTextAtOffset(0, BOUNDARY_LINE_END, "hello my friend", 0, 15,
		       "input", kOk, kOk, kOk,
		       "div", kOk, kOk, kOk,
		       "editable", kOk, kOk, kOk,
		       "textarea", kOk, kOk, kOk);
      testTextAtOffset(1, BOUNDARY_LINE_END, "hello my friend", 0, 15,
		       "input", kOk, kOk, kOk,
		       "div", kOk, kOk, kOk,
		       "editable", kOk, kOk, kOk,
		       "textarea", kOk, kOk, kOk);
      testTextAtOffset(14, BOUNDARY_LINE_END, "hello my friend", 0, 15,
		       "input", kOk, kOk, kOk,
		       "div", kOk, kOk, kOk,
		       "editable", kOk, kOk, kOk,
		       "textarea", kOk, kOk, kOk);
      testTextAtOffset(15, BOUNDARY_LINE_END, "hello my friend", 0, 15,
		       "input", kTodo, kOk, kTodo,
		       "div", kTodo, kOk, kTodo,
		       "editable", kTodo, kOk, kTodo,
		       "textarea", kTodo, kOk, kTodo);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="nsIAccessibleText getText related function tests for html:input,html:div and html:textarea"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=452769">
    Bug 452769
  </a>
  <a target="_blank"
     title="getTextAtOffset for word boundaries: beginning of a new life"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=853340">
    Bug 853340
  </a>
  <a target="_blank"
     title="getTextBeforeOffset for word boundaries: evolving"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=855732">
    Bug 855732
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <input id="input" value="hello my friend"/>
  <div id="div">hello my friend</div>
  <div id="editable" contenteditable="true">hello my friend</div>
  <textarea id="textarea">hello my friend</textarea>

</body>
</html>
